Login logout configurations, services and temporal test controller
